pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kraft"
version = "0.1.0"
description = "KRAFT training pipeline"
authors = [{ name = "Jimin Lee" }, { name = "Sarang Han" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.0",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "torch>=2.1",
    "ta>=0.11",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "PyYAML>=6.0"
]

[tool.setuptools]
packages = ["kraft"]

[project.optional-dependencies]
dev = ["ipykernel", "jupyter", "pytest", "ruff"]

[project.scripts]
kraft-train = "kraft.runner.cli.train:main"

[dependency-groups]
dev = ["ipykernel", "jupyter", "pytest", "ruff"]
